requirements.txt
# EPAS Agent - UiPath Challenge Requirements
# Python 3.10+

# Core Framework
#uipath-sdk>=1.0.0
#uipath>=1.0.0

# LangChain Ecosystem
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-core>=0.1.0

# PDF Processing
pypdf>=3.17.0
pypdf2>=3.0.1

# Vector Store & Embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# OpenAI
openai>=1.10.0
tiktoken>=0.5.2

# API & Server
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0

# Optional: Alternative Vector Stores
chromadb>=0.4.22

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
black>=24.1.0
flake8>=7.0.0

# Logging & Monitoring
loguru>=0.7.2
